backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
web3==6.15.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
